I've created a Behationary/Config class to handle configuration of the tool.
The old config.php just contained a function to retrieve the context definitions, but actually I think we're going to need a bit more configuration than this.  Encapsulating the configuration in a Config class makes it easier to manage this in a sane way, so changes to the way configuration is handled remain transparent to the rest of the code.

Currently the Config class retains the old behaviour, but I plan to extend it shortly.