Fix to the Config class.  The getContexts() function in the local config.php may require functionality that is not available at the time the file is loaded, so we defer calling it until the contexts are actually requested.
I don't think this makes a difference in the current code-base, but I have some work-in-progress that changes the load order to put the config file earlier in the process, at a point where the autoloader is not fully set up.  In this situation, attempting to load the project contexts immediately would fail as the Behat contexts that they extend cannot be found.  With the new approach this problem is eliminated as we decouple the loading of the config file from the initialisation of the context classes.