I've updated the Config class so it now works as a singleton instance.
This allows us to define the path to the config file in one place (bootstrap.php) and then get a copy of the Config object from anywhere, using Config::get().  This means we no longer have to awkwardly pass around the $configFile variable in api.php, or define it in multiple places if we find there are other entry points that need access to it.