I've extended the Config class so that it is now possible to specify a local Behat installation to use.
In the old code, the local Behat installation was always used, but your project contexts may be based on a different Behat version to the one provided by Behationary.  Now you can set a $behatRootPath variable in your config.php to specify an alternative path to a local Behat installation.