The next step of development is to make Behationary project-aware, i.e. so multiple projects can be served from the same Behationary instance.  I haven't fully decided how we will do this - it might be by allowing for multiple config.php files, or it might be done by switching to a YAML-based config file.
Before worrying about that, though, I want to get the code and interface into a state that it is project-aware.  This will allow me to experiment with formats a bit more easily.

This commit is the first step, which adds static functions to the Config class to allow access to the project list and to get/set the currently selected project.  This model may change is based on the idea that there will be a single config file with multiple projects in it, but that may change.