I've updated Config::getSelectedProject() so that if no project has been selected it now returns the projectId of the first project in the list of available projects.
Previously, it always returned the string "default".  This works fine currently, as we only have one project, and this is its ID, but as soon as multiple projects are available this will no longer necessarily be the case and so a more sensible default of 'first defined project' is required.