I've updated the UI so if there are multiple projects available, it now shows a drop-down list allowing you to select which one to search for.  It defaults to the first project in the list, and we use knockout.js to automatically update the page contents whenever this is changed.
(Note that if only one project is available, the projectId is placed in a hidden input instead (so knockout.js continues to work) and no selector is displayed - just the project name, as before.)

In terms of the back-end, the api script has been updated to accept the projectId as the first element in the URL (i.e. instead of /api/steps/etc  it is now /api/projectId/steps/etc).  We capture the projectId but we don't currently do anything with it (it is ignored).  It looks like a lot of the file has changed, but actually this is just because a lot of lines have been indented an extra level.